pyproject.toml
[tool.poetry]
name = "local-rag-assistant"
version = "0.1.0"
description = "A local-first RAG assistant for knowledge management"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
python = "^3.10"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.4.0"
pydantic-settings = "^2.0.0"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"
jinja2 = "^3.1.2"

# ML/NLP
sentence-transformers = "^2.2.2"
faiss-cpu = "^1.7.4"
torch = "^2.1.0"
transformers = "^4.35.0"
rank-bm25 = "^0.2.2"
langdetect = "^1.0.9"
datasketch = "^1.6.4"

# Document parsing
pypdf = "^3.17.0"
pdfminer-six = "^20231228"
python-docx = "^1.1.0"
pandas = "^2.1.0"
chardet = "^5.2.0"

# Storage
sqlalchemy = "^2.0.0"
alembic = "^1.12.0"

# File watching
watchdog = "^3.0.0"

# CLI
click = "^8.1.7"
rich = "^13.7.0"

# API clients (optional)
openai = "^1.3.0"
httpx = "^0.25.0"

# Utilities
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^4.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.0"
black = "^23.10.0"
isort = "^5.12.0"
ruff = "^0.1.0"
mypy = "^1.6.0"
ipython = "^8.17.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "C90", "I", "N", "B", "UP"]
ignore = ["E501", "B008"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"